requirements.txt
chromadb
langchain-community
langchain-core
langchain-openai
langchain-mcp-adapters
langgraph
langgraph-cli[inmem]
langgraph-checkpoint-sqlite
langgraph-sdk
langgraph-supervisor
langsmith
notebook
openevals
pandas
python-dotenv
slack_sdk
tavily-python
wikipedia
mcp
fastapi
uvicorn
requests